api/requirements.txt
django
djangorestframework
pytz
markdown
mysqlclient
mysql-connector-python
requests
django-cors-headers
drf_yasg
